# backend-go/internal/database/queries/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package database

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Profile struct {
	ID               pgtype.UUID      `json:"id"`
	Email            pgtype.Text      `json:"email"`
	FullName         pgtype.Text      `json:"full_name"`
	AvatarUrl        pgtype.Text      `json:"avatar_url"`
	Plan             pgtype.Text      `json:"plan"`
	StripeCustomerID pgtype.Text      `json:"stripe_customer_id"`
	TeamID           pgtype.UUID      `json:"team_id"`
	CreatedAt        pgtype.Timestamp `json:"created_at"`
	UpdatedAt        pgtype.Timestamp `json:"updated_at"`
}

type Project struct {
	ID              pgtype.UUID      `json:"id"`
	Name            string           `json:"name"`
	Description     pgtype.Text      `json:"description"`
	Slug            string           `json:"slug"`
	OwnerID         pgtype.UUID      `json:"owner_id"`
	DefaultLanguage pgtype.Text      `json:"default_language"`
	Languages       []string         `json:"languages"`
	CreatedAt       pgtype.Timestamp `json:"created_at"`
	UpdatedAt       pgtype.Timestamp `json:"updated_at"`
	IsDeleted       bool             `json:"is_deleted"`
	DeletedAt       pgtype.Timestamp `json:"deleted_at"`
	Archived        bool             `json:"archived"`
}

type Team struct {
	ID        pgtype.UUID      `json:"id"`
	Name      string           `json:"name"`
	OwnerID   pgtype.UUID      `json:"owner_id"`
	CreatedAt pgtype.Timestamp `json:"created_at"`
	UpdatedAt pgtype.Timestamp `json:"updated_at"`
	AvatarUrl pgtype.Text      `json:"avatar_url"`
	Personal  bool             `json:"personal"`
}

type TeamMember struct {
	ID        pgtype.UUID      `json:"id"`
	ProjectID pgtype.UUID      `json:"project_id"`
	UserID    pgtype.UUID      `json:"user_id"`
	Role      string           `json:"role"`
	InvitedAt pgtype.Timestamp `json:"invited_at"`
	JoinedAt  pgtype.Timestamp `json:"joined_at"`
	InvitedBy pgtype.UUID      `json:"invited_by"`
	CreatedAt pgtype.Timestamp `json:"created_at"`
}

type TeamMembership struct {
	ID        pgtype.UUID      `json:"id"`
	TeamID    pgtype.UUID      `json:"team_id"`
	UserID    pgtype.UUID      `json:"user_id"`
	Role      string           `json:"role"`
	JoinedAt  pgtype.Timestamp `json:"joined_at"`
	CreatedAt pgtype.Timestamp `json:"created_at"`
}